Fix R3F Button-in-THREE crash and 404, split DOM HUDs out of Canvas
- Move CameraHUD and MobileHUD DOM rendering outside <Canvas> into App.jsx
  so R3F's reconciler never encounters <button>/<div> elements
- FreeCameraControls/MobileControls now return only Three.js primitives;
  camera state (mode/setMode/plcLockRef) owned by App and passed as props
- Fix 404: EmulatorV86 cdrom URL corrected to TinyCore-current.iso
- Add CLAUDE.md with architecture rules, especially the R3F DOM rendering constraint
