+++ 3D_portfolio/src/App.jsx
@@ -1,5 +1,5 @@
 import { Canvas } from '@react-three/fiber';
-import { Suspense, useState } from 'react';
+import { Suspense, useState, useRef } from 'react';
 import { useProgress, useGLTF } from '@react-three/drei';
 import RoomScene from './models/RoomScene';
 import VTuberView from './components/VTuberView';
@@ -8,20 +8,24 @@ import BackgroundMusic from './components/BackgroundMusic';
 import Loader from './components/Loader';
 import FreeCameraControls from './components/FreeCameraControls';
 import MobileControls from './components/MobileControls';
-
+import CameraHUD from './components/CameraHUD';
+import MobileHUD from './components/MobileHUD';
 
 useGLTF.preload('/models/VTuber2.vrm');
 
 export default function App() {
   const { progress } = useProgress();
-  const [showLoader, setShowLoader] = useState(true);
+  const [showLoader, setShowLoader]   = useState(true);
   const [sceneVisible, setSceneVisible] = useState(false);
-  const [playMusic, setPlayMusic] = useState(false);
+  const [playMusic, setPlayMusic]     = useState(false);
+
+  // Camera mode owned here so HUDs (outside Canvas) and controls (inside Canvas) share it
+  const [cameraMode, setCameraMode] = useState('preset');
+  const plcLockRef = useRef(null); // set by FreeCameraControls, called by CameraHUD
 
   const handleEnter = () => {
     const loaderEl = document.getElementById('loading-overlay');
     if (loaderEl) loaderEl.classList.add('fade-out');
-
     setTimeout(() => {
       setShowLoader(false);
       setSceneVisible(true);
@@ -31,10 +35,7 @@ export default function App() {
 
   return (
     <>
-      {showLoader && (
-        <Loader progress={progress} onEnter={handleEnter} />
-      )}
-
+      {showLoader && <Loader progress={progress} onEnter={handleEnter} />}
       {playMusic && <BackgroundMusic />}
 
       <Canvas
@@ -56,9 +57,35 @@ export default function App() {
           </Suspense>
         )}
 
-        {sceneVisible && <FreeCameraControls />}
-        {sceneVisible && <MobileControls />}
+        {sceneVisible && (
+          <FreeCameraControls
+            mode={cameraMode}
+            setMode={setCameraMode}
+            plcLockRef={plcLockRef}
+          />
+        )}
+        {sceneVisible && (
+          <MobileControls
+            mode={cameraMode}
+            setMode={setCameraMode}
+          />
+        )}
       </Canvas>
+
+      {/* HUDs rendered outside Canvas — safe DOM territory */}
+      {sceneVisible && (
+        <CameraHUD
+          mode={cameraMode}
+          setMode={setCameraMode}
+          plcLockRef={plcLockRef}
+        />
+      )}
+      {sceneVisible && (
+        <MobileHUD
+          mode={cameraMode}
+          setMode={setCameraMode}
+        />
+      )}
     </>
   );
 }
