+++ 3D_portfolio/src/components/FreeCameraControls.jsx
@@ -1,29 +1,14 @@
 // src/components/FreeCameraControls.jsx
+// Three.js / R3F logic only — NO DOM rendering.
+// DOM HUD lives in CameraHUD.jsx, rendered outside <Canvas> in App.jsx.
 import { useThree, useFrame } from '@react-three/fiber';
 import { PointerLockControls } from '@react-three/drei';
-import { useEffect, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
+import { useEffect, useRef } from 'react';
 import * as THREE from 'three';
 
-// ─────────────────────────────────────────────────────────────────────────────
-// SPAWN POINT — camera starts here when the site loads.
-//
-// Room world bounds (from GLB): X 1.67–3.76  Y 3.19–4.19  Z 2.68–4.89
-// Monitors sit at Z ≈ 2.82.  Bed is at the far / high-Z end.
-//
-// POSITION: above-bed area, slightly elevated.
-// LOOK_AT:  toward the monitor cluster.
-//
-// To adjust: move in-game, then open browser console and type:
-//   window.__camera.position  → copy those values here
-// ─────────────────────────────────────────────────────────────────────────────
-const SPAWN_POSITION = new THREE.Vector3(2.7, 4.05, 4.6);
-const SPAWN_LOOK_AT  = new THREE.Vector3(2.9, 3.55, 2.85);
+export const SPAWN_POSITION = new THREE.Vector3(2.7, 4.05, 4.6);
+export const SPAWN_LOOK_AT  = new THREE.Vector3(2.9, 3.55, 2.85);
 
-// ─────────────────────────────────────────────────────────────────────────────
-// CAMERA PRESETS — the buttons users click.
-// Adjust positions once you've explored the room.
-// ─────────────────────────────────────────────────────────────────────────────
 export const CAMERA_PRESETS = [
   {
     name: 'Overview',
@@ -47,16 +32,11 @@ export const CAMERA_PRESETS = [
   },
 ];
 
-// Expose camera globally so users can find spawn coords in the console
 if (typeof window !== 'undefined') window.__cameraPresets = CAMERA_PRESETS;
 
-// ─────────────────────────────────────────────────────────────────────────────
+const _lookDir = new THREE.Matrix4();
 
-const _targetPos  = new THREE.Vector3();
-const _targetQuat = new THREE.Quaternion();
-const _lookDir    = new THREE.Matrix4();
-
-function quatFromLookAt(position, target) {
+export function quatFromLookAt(position, target) {
   _lookDir.lookAt(position, target, new THREE.Vector3(0, 1, 0));
   const q = new THREE.Quaternion();
   q.setFromRotationMatrix(_lookDir);
@@ -64,225 +44,65 @@ function quatFromLookAt(position, target) {
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
-// HUD — rendered via portal into document.body
+// Props:
+//   mode        — 'preset' | 'confirming' | 'free'  (owned by App)
+//   setMode     — setter from App
+//   plcLockRef  — ref whose .current will be set to () => plcRef.current.lock()
+//                 so CameraHUD can trigger pointer lock without being in Canvas
 // ─────────────────────────────────────────────────────────────────────────────
-function CameraHUD({ mode, onRequestFree, onConfirmFree, onCancelFree, onExitFree }) {
-  // ── Confirmation overlay ──
-  if (mode === 'confirming') {
-    return createPortal(
-      <div style={{
-        position: 'fixed', inset: 0,
-        background: 'rgba(0,0,0,0.75)',
-        backdropFilter: 'blur(4px)',
-        display: 'flex', alignItems: 'center', justifyContent: 'center',
-        zIndex: 10000,
-      }}>
-        <div style={{
-          background: '#0f172a',
-          border: '1px solid rgba(255,255,255,0.15)',
-          borderRadius: 12,
-          padding: '28px 32px',
-          maxWidth: 380,
-          width: '90vw',
-          color: '#fff',
-          fontFamily: 'system-ui, sans-serif',
-          boxShadow: '0 25px 60px rgba(0,0,0,0.6)',
-        }}>
-          <div style={{ fontSize: 20, fontWeight: 700, marginBottom: 8 }}>
-            🔓 Enable Free Camera?
-          </div>
-          <div style={{ fontSize: 13, color: '#94a3b8', lineHeight: 1.6, marginBottom: 20 }}>
-            This will lock your mouse cursor to the window for first-person exploration.
-          </div>
-          <div style={{
-            background: '#1e293b', borderRadius: 8, padding: '12px 16px',
-            marginBottom: 20, fontSize: 12, color: '#cbd5e1', lineHeight: 2,
-          }}>
-            <div><kbd style={kbdStyle}>W A S D</kbd> &nbsp;Move</div>
-            <div><kbd style={kbdStyle}>Mouse</kbd> &nbsp;Look around</div>
-            <div><kbd style={kbdStyle}>Space</kbd> / <kbd style={kbdStyle}>Ctrl</kbd> &nbsp;Up / Down</div>
-            <div><kbd style={kbdStyle}>Shift</kbd> &nbsp;Sprint</div>
-            <div><kbd style={kbdStyle}>Esc</kbd> &nbsp;Exit free camera</div>
-          </div>
-          <div style={{ display: 'flex', gap: 10 }}>
-            <button onClick={onCancelFree} style={btnSecondary}>Cancel</button>
-            <button onClick={onConfirmFree} style={btnPrimary}>Enter Free Camera</button>
-          </div>
-        </div>
-      </div>,
-      document.body
-    );
-  }
-
-  // ── Free camera mode — persistent tutorial banner + exit button ──
-  if (mode === 'free') {
-    return createPortal(
-      <div style={{
-        position: 'fixed', bottom: 0, left: 0, right: 0,
-        background: 'rgba(0,0,0,0.7)',
-        backdropFilter: 'blur(8px)',
-        borderTop: '1px solid rgba(255,255,255,0.1)',
-        padding: '8px 16px',
-        display: 'flex', alignItems: 'center', justifyContent: 'space-between',
-        zIndex: 9999, fontFamily: 'system-ui, sans-serif',
-        flexWrap: 'wrap', gap: 8,
-      }}>
-        <div style={{ display: 'flex', alignItems: 'center', gap: 16, flexWrap: 'wrap' }}>
-          <span style={{ color: '#64748b', fontSize: 11 }}>FREE CAMERA</span>
-          {[
-            ['W A S D', 'Move'],
-            ['Mouse', 'Look'],
-            ['Space / Ctrl', 'Up / Down'],
-            ['Shift', 'Sprint'],
-            ['Esc', 'Exit'],
-          ].map(([key, label]) => (
-            <span key={key} style={{ fontSize: 11, color: '#94a3b8' }}>
-              <kbd style={kbdStyle}>{key}</kbd> {label}
-            </span>
-          ))}
-        </div>
-        <button onClick={onExitFree} style={btnSecondary}>
-          ✕ Exit Free Camera
-        </button>
-      </div>,
-      document.body
-    );
-  }
-
-  // ── Default preset mode — navigation buttons ──
-  return createPortal(
-    <div style={{
-      position: 'fixed', bottom: 20, left: '50%',
-      transform: 'translateX(-50%)',
-      display: 'flex', gap: 8, alignItems: 'center',
-      zIndex: 9999, fontFamily: 'system-ui, sans-serif',
-      flexWrap: 'wrap', justifyContent: 'center',
-      maxWidth: '95vw',
-    }}>
-      {/* Preset buttons */}
-      {CAMERA_PRESETS.map((p, i) => (
-        <button
-          key={p.name}
-          data-preset={i}
-          style={{
-            ...btnPreset,
-            // Active state via dataset (set in useFrame below)
-          }}
-          title={`Go to: ${p.name}`}
-        >
-          {p.name}
-        </button>
-      ))}
-
-      {/* Divider */}
-      <div style={{ width: 1, height: 24, background: 'rgba(255,255,255,0.15)' }} />
-
-      {/* Free camera unlock */}
-      <button onClick={onRequestFree} style={btnUnlock}>
-        🔓 Free Camera
-      </button>
-    </div>,
-    document.body
-  );
-}
-
-// ── Inline styles ──────────────────────────────────────────────────────────
-const kbdStyle = {
-  background: '#1e293b', border: '1px solid #334155',
-  borderRadius: 4, padding: '1px 6px', fontSize: 11,
-  fontFamily: 'monospace', color: '#e2e8f0',
-};
-const btnBase = {
-  border: 'none', cursor: 'pointer', borderRadius: 8,
-  fontFamily: 'system-ui, sans-serif', fontWeight: 500,
-  transition: 'all 0.15s',
-};
-const btnPrimary = {
-  ...btnBase, flex: 1,
-  background: '#3b82f6', color: '#fff',
-  padding: '9px 16px', fontSize: 13,
-};
-const btnSecondary = {
-  ...btnBase,
-  background: 'rgba(255,255,255,0.08)',
-  color: '#94a3b8',
-  border: '1px solid rgba(255,255,255,0.12)',
-  padding: '7px 14px', fontSize: 12,
-};
-const btnPreset = {
-  ...btnBase,
-  background: 'rgba(15,23,42,0.85)',
-  color: '#e2e8f0',
-  border: '1px solid rgba(255,255,255,0.15)',
-  backdropFilter: 'blur(8px)',
-  padding: '8px 16px', fontSize: 12,
-};
-const btnUnlock = {
-  ...btnBase,
-  background: 'rgba(59,130,246,0.15)',
-  color: '#93c5fd',
-  border: '1px solid rgba(59,130,246,0.3)',
-  backdropFilter: 'blur(8px)',
-  padding: '8px 16px', fontSize: 12,
-};
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Main component
-// ─────────────────────────────────────────────────────────────────────────────
-export default function FreeCameraControls() {
+export default function FreeCameraControls({ mode, setMode, plcLockRef }) {
   const { camera } = useThree();
-  const [mode, setMode] = useState('preset'); // 'preset' | 'confirming' | 'free'
   const plcRef = useRef();
 
-  const presetTarget = useRef(null);  // { position, quaternion }
+  const presetTarget = useRef(null);
   const lerpingRef   = useRef(false);
-  const activePreset = useRef(0);
 
-  const keys = useRef({});
+  const keys      = useRef({});
   const direction = useRef(new THREE.Vector3());
 
-  // ── Expose camera globally for coordinate discovery ─────────────────────
-  useEffect(() => {
-    window.__camera = camera;
-  }, [camera]);
+  // Expose camera globally for coordinate discovery
+  useEffect(() => { window.__camera = camera; }, [camera]);
 
-  // ── Set spawn point and rotation order ──────────────────────────────────
+  // Expose pointer-lock trigger to CameraHUD (outside Canvas)
+  useEffect(() => {
+    if (plcLockRef) plcLockRef.current = () => plcRef.current?.lock?.();
+  }, [plcLockRef]);
+
+  // Spawn position + rotation order
   useEffect(() => {
     camera.rotation.order = 'YXZ';
     camera.position.copy(SPAWN_POSITION);
     camera.quaternion.copy(quatFromLookAt(SPAWN_POSITION, SPAWN_LOOK_AT));
   }, [camera]);
 
-  // ── Wire preset buttons (event delegation on the HUD) ───────────────────
+  // Unlock pointer when leaving free mode
+  useEffect(() => {
+    if (mode !== 'free') plcRef.current?.unlock?.();
+  }, [mode]);
+
+  // Preset button delegation — CameraHUD buttons carry data-preset attribute
   useEffect(() => {
     const handler = (e) => {
       const btn = e.target.closest('[data-preset]');
       if (!btn) return;
-      const idx = parseInt(btn.dataset.preset, 10);
-      goToPreset(idx);
+      goToPreset(parseInt(btn.dataset.preset, 10));
     };
     document.addEventListener('click', handler);
     return () => document.removeEventListener('click', handler);
-  }, []);
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
-  // ── Keyboard ─────────────────────────────────────────────────────────────
+  // Keyboard
   useEffect(() => {
-    const dn = (e) => {
-      keys.current[e.code] = true;
-      if (e.code === 'Escape' && mode === 'free') {
-        exitFree();
-      }
-    };
+    const dn = (e) => { keys.current[e.code] = true; };
     const up = (e) => { keys.current[e.code] = false; };
     window.addEventListener('keydown', dn);
-    window.addEventListener('keyup', up);
+    window.addEventListener('keyup',   up);
     return () => {
       window.removeEventListener('keydown', dn);
-      window.removeEventListener('keyup', up);
+      window.removeEventListener('keyup',   up);
     };
-  }, [mode]);
+  }, []);
 
-  // ── Mode transitions ──────────────────────────────────────────────────────
   const goToPreset = (idx) => {
     const p = CAMERA_PRESETS[idx];
     if (!p) return;
@@ -291,27 +111,10 @@ export default function FreeCameraControls() {
       quaternion: quatFromLookAt(p.position, p.lookAt),
     };
     lerpingRef.current = true;
-    activePreset.current = idx;
   };
 
-  const requestFree  = () => setMode('confirming');
-  const cancelFree   = () => setMode('preset');
-
-  const confirmFree  = () => {
-    setMode('free');
-    // Small delay so the confirm button click doesn't interfere with pointer lock
-    setTimeout(() => plcRef.current?.lock?.(), 150);
-  };
-
-  const exitFree = () => {
-    plcRef.current?.unlock?.();
-    setMode('preset');
-  };
-
-  // ── Per-frame ─────────────────────────────────────────────────────────────
   useFrame((_, delta) => {
     if (mode === 'free') {
-      // ── Free camera WASD ──
       const speed = keys.current['ShiftLeft'] ? 15 : 5;
       direction.current.set(0, 0, 0);
       if (keys.current['KeyW']) direction.current.z += 1;
@@ -331,7 +134,6 @@ export default function FreeCameraControls() {
       return;
     }
 
-    // ── Preset lerp (preset mode) ──
     if (lerpingRef.current && presetTarget.current) {
       const t = Math.min(delta * 5, 1);
       camera.position.lerp(presetTarget.current.position, t);
@@ -344,21 +146,7 @@ export default function FreeCameraControls() {
     }
   });
 
-  return (
-    <>
-      {mode === 'free' && (
-        <PointerLockControls
-          ref={plcRef}
-          onUnlock={exitFree}
-        />
-      )}
-      <CameraHUD
-        mode={mode}
-        onRequestFree={requestFree}
-        onConfirmFree={confirmFree}
-        onCancelFree={cancelFree}
-        onExitFree={exitFree}
-      />
-    </>
-  );
+  return mode === 'free' ? (
+    <PointerLockControls ref={plcRef} onUnlock={() => setMode('preset')} />
+  ) : null;
 }
